add next_element parameter
next_element parameter holds the next element when a particle
intersects a face. This is added to both the trace_particle_through_mesh
and the functor. The functor may or may not use the value, but there's
a high chance to be used.

[Important] The test is failing. This is because although the particles
doesn't intersect face 0 (it is along the way of the ray going forward)
the closeness algorithm finds it as the intersected face.